% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uconn.R
\name{uconn}
\alias{uconn}
\title{UCONN - Unconstrained/Constrained Ordination Neural Network}
\usage{
uconn(
  Y,
  X = NULL,
  iterations = 200,
  learning_rate = 0.001,
  l2 = 0,
  architecture = c(50, 20, 10),
  latent_dim = 2L,
  response = c("binary", "regression"),
  KLD_weight = 1
)
}
\arguments{
\item{Y}{Community matrix, columns correspond to the species}

\item{X}{A matrix or data frame of env variables. If `NULL`, only `Y` will be used as input. Factors will be one-hot encoded.}

\item{iterations}{An integer specifying the number of training iterations. Default is 200.}

\item{learning_rate}{A numeric value specifying the learning rate for model optimization. Default is 0.001.}

\item{l2}{A numeric value specifying the L2 regularization penalty. Default is 0.0 (no regularization).}

\item{architecture}{A numeric vector specifying the number of units in each hidden layer of the model. Default is `c(50, 20, 10)`.}

\item{latent_dim}{An integer specifying the dimension of the latent space. Default is 2.}

\item{response}{A character string specifying the response type. Possible values are `"binary"` or `"regression"` for continuous outcomes. Default is `"binary"`.}

\item{KLD_weight}{A numeric value specifying the weight of the Kullback-Leibler divergence term in the loss function. Default is 5.0.}
}
\value{
An object of class `uconn` containing:
  \item{latent}{A matrix of latent representations of the data in the specified latent dimension.}
  \item{importance}{A matrix or data frame indicating the importance of the covariates (if `X` is provided).}
  \item{model}{A serialized model object for reuse or further analysis.}
}
\description{
Fits a conditional variational autoencoder (cVAE) model to the given community matrix, optionally constrained/conditionally on covariates
to obtain latent representations that can be interpreted as an ordination. This model can be used for binary (presence-absence) or regression responses.
}
\examples{
n = 100
sp = 10
e = 2
X = mvtnorm::rmvnorm(n, sigma = diag(1.0, e))
w = mvtnorm::rmvnorm(sp, sigma = diag(1.0, e))
sigma = cov2cor(rWishart(1, sp, Sigma = diag(1.0, sp))[,,1])
Y = 1*(X \%*\% t(w) + mvtnorm::rmvnorm(n, sigma = sigma) < 0 )


model = uconn(Y, X)
model
plot(model)


}
